Rename `stalling` mod to `stacks_mining`
