+++ testnet/stacks-node/src/tests/signer/commands/stacks_mining.rs
